+++ CHANGES.rst
@@ -11,6 +11,10 @@ Changelog
 2.1.15 (unreleased)
 -------------------
 
+- Added the missing Authenticated and Anonymous roles within the
+  `RolesVocabulary``.
+  [ichim-david]
+
 - Make ``KeywordsVocabulary`` more customizeable using an ``keyword_index``
   class variable to allow users to inherit and just override that attribute
   to build their own keyword vocabularies.